requirements.txt
bidict==0.22.1
click==8.1.3
Flask==2.2.3
Flask-SocketIO==5.3.3
greenlet==2.0.2
h11==0.14.0
importlib-metadata==6.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-engineio==4.4.0
python-socketio==5.8.0
simple-websocket==0.10.0
SQLAlchemy==2.0.9
typing-extensions==4.5.0
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.15.0
bcrypt
virtualenv
pyOpenSSL
PyJWT